# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timestamp Converter API",
    "description": "API for converting timestamps between formats and timezones with support for batch processing, custom formatting, and visualization.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://tsconv.com/support",
      "email": "support@tsconv.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://tsconv.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/convert": {
      "get": {
        "tags": ["conversion"],
        "summary": "Convert a timestamp or date string",
        "description": "Convert between Unix timestamps and human-readable dates with optional timezone conversion and custom formatting.",
        "operationId": "convertTimestamp",
        "parameters": [
          {
            "name": "timestamp",
            "in": "query",
            "description": "Unix timestamp in seconds",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "ISO 8601 date string",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Custom format name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "in": "query",
            "description": "Source timezone",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetTimezone",
            "in": "query",
            "description": "Target timezone for conversion",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeFormats",
            "in": "query",
            "description": "Include all available formats",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful conversion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enhanced-batch": {
      "post": {
        "tags": ["batch", "conversion"],
        "summary": "Convert multiple timestamps or dates in a single request",
        "description": "Batch convert multiple timestamps or date strings with custom formatting and timezone conversion.",
        "operationId": "batchConvert",
        "requestBody": {
          "description": "Batch conversion request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchConversionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful batch conversion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchConversionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/formats": {
      "get": {
        "tags": ["format"],
        "summary": "Get information about supported date/time formats",
        "description": "Discover and use custom date/time formats.",
        "operationId": "getFormats",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category (standard, human, regional, technical)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["standard", "human", "regional", "technical"]
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Get specific format details",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful format information retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConversionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "input": {
                "type": "string",
                "example": "1640995200"
              },
              "timestamp": {
                "type": "number",
                "example": 1640995200
              },
              "formats": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "example": {
                  "iso8601": "2022-01-01T00:00:00.000Z",
                  "utc": "Sat, 01 Jan 2022 00:00:00 GMT",
                  "local": "12/31/2021, 7:00:00 PM",
                  "custom": "12/31/2021 07:00 PM",
                  "relative": "2 years ago"
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "processingTime": {
                "type": "number",
                "example": 25
              },
              "cache": {
                "type": "object",
                "properties": {
                  "hit": {
                    "type": "boolean",
                    "example": false
                  },
                  "ttl": {
                    "type": "number",
                    "example": 300
                  }
                }
              }
            }
          }
        }
      },
      "BatchConversionRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "example": [1640995200, "2022-01-01", "2022-12-31T23:59:59Z"]
          },
          "outputFormat": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["iso8601", "timestamp", "us-date"]
          },
          "timezone": {
            "type": "string",
            "example": "UTC"
          },
          "targetTimezone": {
            "type": "string",
            "example": "America/New_York"
          },
          "options": {
            "type": "object",
            "properties": {
              "continueOnError": {
                "type": "boolean",
                "example": true
              },
              "maxItems": {
                "type": "number",
                "example": 100
              }
            }
          }
        }
      },
      "BatchConversionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "input": {
                  "oneOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "string"
                    }
                  ],
                  "example": 1640995200
                },
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "number",
                      "example": 1640995200
                    },
                    "formats": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      },
                      "example": {
                        "iso8601": "2022-01-01T00:00:00.000Z",
                        "timestamp": "1640995200",
                        "us-date": "12/31/2021"
                      }
                    }
                  }
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "VALIDATION_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid timestamp format"
                    }
                  }
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "number",
                "example": 3
              },
              "successCount": {
                "type": "number",
                "example": 3
              },
              "errorCount": {
                "type": "number",
                "example": 0
              },
              "processingTime": {
                "type": "number",
                "example": 120
              }
            }
          }
        }
      },
...
      "FormatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "category": {
                "type": "string",
                "example": "regional"
              },
              "formats": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "US Date"
                    },
                    "pattern": {
                      "type": "string",
                      "example": "MM/DD/YYYY"
                    },
                    "description": {
                      "type": "string",
                      "example": "US date format"
                    },
                    "example": {
                      "type": "string",
                      "example": "01/15/2024"
                    },
                    "category": {
                      "type": "string",
                      "example": "regional"
                    }
                  }
                }
              },
              "count": {
                "type": "number",
                "example": 5
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "Invalid timestamp format"
              },
              "details": {
                "type": "object",
                "properties": {
                  "suggestions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["Use Unix timestamp format"]
                  }
                }
              },
              "timestamp": {
                "type": "string",
                "example": "2024-01-15T12:00:00.000Z"
              }
            }
          }
        }
      }
    }
  }
}